LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-style-changes-001-expected.txt


FAIL .before 1 assert_equals: 
<div id="item3" data-expected-width="50" data-expected-height="30" data-offset-x="100" class=" before" data-offset-y="20"></div>
offsetTop expected 20 but got 0
PASS .before 2 
FAIL .before 3 assert_equals: 
<div id="item1" data-expected-width="50" data-expected-height="20" data-offset-x="0" class=" before" data-offset-y="30"></div>
offsetTop expected 30 but got 0
